Move AP appointment scheduling to new controller
- Create a new `AppointmentsController` which will be responsible for
Appointment/Session related things.
- Move first stage of scheduling an Action Plan appointment to new
controller, and remove from the over-used
`ServiceProviderReferralsController`
- Rename `*Session`-based functions to make it clear that they're
actually referring to `Appointments` to hopefully clarify that
`Appointment` !== `Session`
- Refactor overly-DRY controller functions as they were fairly hard to
read and involved a lot of scrolling up and down the file and passing
functions around.
